+++ RELEASE_NOTES.md
@@ -23,6 +23,7 @@ Version 2.4.0 (ETA - December 2016)
  * exclude build files from bower install (pull request #75)
  * ensure WILDCARD events are included in list of available transitions() (issue #93)
  * fix FSM getting stuck into "*" state when using double wildcard (issue #64)
+ * function (fsm.states) returning list of all available states in the machine would help automated testing (issue #54)
 
 Version 2.3.5 (January 20 2014)
 -------------------------------